n8n_data/requirements.txt
streamlit
fastapi
uvicorn
requests
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv

